requirements_all.txt
# All dependencies for TESTMODELIIDOSTRAJANIE

# OCR & Vision
google-cloud-vision>=3.4.0

# LLM APIs
openai>=1.3.0
ollama>=0.1.0

# Data & Analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.27.3
Pillow>=9.5.0
pdf2image>=1.16.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
